+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/configuration/components/PreviewPanel.kt
@@ -23,6 +23,7 @@ internal class PreviewPanel(
     private val showFactionCrests: () -> Boolean,
     private val sameVehicleVelocity: () -> Boolean,
     private val enableNewVehicles: () -> Boolean,
+    private val solidProgressBarColor: () -> Boolean,
     private val enabledVehicles: () -> Map<String, Boolean>?
 ) : JTitledPanel(StarWarsBundle.message(BundleConstants.PREVIEW_TITLE)) {
 
@@ -85,7 +86,8 @@ internal class PreviewPanel(
                 showVehicleNames,
                 showToolTips,
                 showFactionCrests,
-                sameVehicleVelocity
+                sameVehicleVelocity,
+                solidProgressBarColor
             )
         )
 
@@ -95,7 +97,8 @@ internal class PreviewPanel(
                 showVehicleNames,
                 showToolTips,
                 showFactionCrests,
-                sameVehicleVelocity
+                sameVehicleVelocity,
+                solidProgressBarColor
             )
         )
     }