+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/constants/ConfigurationConstants.kt
@@ -5,3 +5,4 @@ internal const val DEFAULT_SHOW_TOOLTIPS = true
 internal const val DEFAULT_SHOW_FACTION_CRESTS = false
 internal const val DEFAULT_SAME_VEHICLE_VELOCITY = false
 internal const val DEFAULT_ENABLE_NEW_VEHICLES = true
+internal const val DEFAULT_SOLID_PROGRESS_BAR_COLOR = false